Fix formatting issues + Refactor analytics components
