+++ modules/react-components/src/analytics/index.ts
@@ -16,7 +16,6 @@
  * under the License.
  */
 
-export * from "./stat-count-card";
 export * from "./stats-overview-widget";
 export * from "./stats-insights-widget";
 export * from "./stats-quick-links-widget";
